+++ tests/Regressors/RegressionTreeTest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Tests\Regressors;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Verbose;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Graph\Trees\CART;
@@ -35,6 +36,7 @@ class RegressionTreeTest extends TestCase
         $this->assertInstanceOf(RegressionTree::class, $this->estimator);
         $this->assertInstanceOf(CART::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
+        $this->assertInstanceOf(Verbose::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }