+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - t-SNE embedder is now Verbose
     - Fixed Naive Bayes reset category counts during partial train
     - Pipeline and Persistent Model are now Verbose
+    - Classification and Regression trees now Verbose
 
 - 0.0.6-alpha
     - Added Gradient Boost regressor